requirements.txt
numpy
pandas
Pillow
streamlit
zipfile39
tf-keras
uform
torch 
torchvision
hf_xet
deepface